LARGE REWORK WARNING - Please code review this commit (there are still kinks in Model Binding)
Modified - Changed Choices from a string to be its own Entity and table. This was done such that its more extensible and each choice may now have an associated image.

Modified - in doing so the creation / edit pages were changed. A drop down list of available icons from the Content/images/choiceImage folder was added.

Removed - Now that Choices became more extensible there wasn't need for some code that was breaking the DRY principle (having Selctions, Options, And Choices was really dumb).

IMPORTANT NOTE - I normally do not commit code that is not thoroughly tested but enough progress was made that I do not want to lose the changes and it is showing promise.